backend/doc_ingestor/requirements.txt
# =============================================================================
# DEPENDENCIES - Python packages
# =============================================================================

# FastAPI et serveur
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Rate Limiting (Protection DDoS)
slowapi==0.1.9

# Monitoring et Métriques
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0

# Base de données
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0  # Driver async PostgreSQL
psycopg2-binary==2.9.9  # Gardé pour Alembic
alembic==1.13.1

# RabbitMQ
aio-pika==9.3.1  # Client async RabbitMQ
pika==1.3.2  # Gardé pour compatibilité

# Configuration
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.1

# Extraction de contenu
tika==2.6.0
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.2.0

# Utilitaires
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Logging et monitoring
python-json-logger==2.0.7

# HTTP client
httpx==0.26.0
aiofiles==23.2.1

# Tests
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Resilience & Retry
tenacity==8.2.3  # Retry avec exponential backoff
pybreaker==1.0.1  # Circuit breaker pattern

